Refactored and fixed imgui integration and added UI to demo 068_AO_WS

+++ framework/imgui/imgui.h
@@ -1130,7 +1130,7 @@ struct ImDrawCmd
 
 // Vertex index (override with '#define ImDrawIdx unsigned int' inside in imconfig.h)
 #ifndef ImDrawIdx
-typedef unsigned short ImDrawIdx;
+typedef unsigned int ImDrawIdx;
 #endif
 
 // Vertex layout
